+++ source/common/http/conn_manager_utility.cc
@@ -77,6 +77,14 @@ Network::Address::InstanceConstSharedPtr ConnectionManagerUtility::mutateRequest
     }
     request_headers.insertForwardedProto().value().setReference(
         connection.ssl() ? Headers::get().SchemeValues.Https : Headers::get().SchemeValues.Http);
+
+    // Since useRemoteAddress() is true, skip adding x-forwarded-port when the x-forwarded-port
+    // is already set.
+    if (config.appendXForwardedPort() && !request_headers.ForwardedPort()) {
+      // Otherwise, set x-forwarded-for with remote address connection port, i.e. the port of the
+      // proxy in front of this Envoy that is used by the client to connect to.
+      request_headers.insertForwardedPort().value(connection.remoteAddress()->ip()->port());
+    }
   } else {
     // If we are not using remote address, attempt to pull a valid IPv4 or IPv6 address out of XFF.
     // If we find one, it will be used as the downstream address for logging. It may or may not be
@@ -84,6 +92,12 @@ Network::Address::InstanceConstSharedPtr ConnectionManagerUtility::mutateRequest
     auto ret = Utility::getLastAddressFromXFF(request_headers, xff_num_trusted_hops);
     final_remote_address = ret.address_;
     single_xff_address = ret.single_address_;
+
+    if (config.appendXForwardedPort()) {
+      // Since we are not using remote address, we set x-forwarded-port with local connection port
+      // (i.e. the port that is used by the client to connect to this Envoy).
+      request_headers.insertForwardedPort().value(connection.localAddress()->ip()->port());
+    }
   }
 
   // If we didn't already replace x-forwarded-proto because we are using the remote address, and